Add tracking apparatus
Now modules take a "tracker" string when they are created. Every module
also has a "log" method. Calling model.log() will propagate the log
instruction to every leaf module. An example use case is to track the
weight norm during a training run. The mechanism for concatenating log
information is to union child dictionaries; each has a top-level key
according to its unique tracker, which is merged into a single
model-level dictionary that is returned to the user.